+++ pyssata/lib/compute_zern_ifunc.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from pyssata import xp
-from pyssata import standard_dtype
+from pyssata import float_dtype
 
 from pyssata.lib.make_mask import make_mask
 from pyssata.lib.zernike_generator import ZernikeGenerator
@@ -20,7 +20,7 @@ def compute_zern_ifunc(dim, nzern, obsratio=0.0, diaratio=1.0, start_mode=0, mas
     zern_phase_3d = zern_phase_3d[start_mode:]
     nzern -= start_mode
 
-    zern_phase_2d = xp.array([zern_phase_3d[i][idx] for i in range(nzern)], dtype=standard_dtype)
+    zern_phase_2d = xp.array([zern_phase_3d[i][idx] for i in range(nzern)], dtype=float_dtype)
     
     zern_phase_2d = zern_phase_2d / xp.std(zern_phase_2d, axis=1, keepdims=True)
